+++ app/controllers/recording_studio_orderable/home_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module RecordingStudioOrderable
+  class HomeController < ApplicationController
+    def index; end
+  end
+end