+++ lib/recording_studio_orderable/engine.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+module RecordingStudioOrderable
+  class Engine < ::Rails::Engine
+    isolate_namespace RecordingStudioOrderable
+
+    initializer "recording_studio_orderable.load_config" do |app|
+      if app.respond_to?(:config_for)
+        yaml = app.config_for(:recording_studio_orderable)
+        RecordingStudioOrderable.configuration.merge!(yaml) if yaml.respond_to?(:each)
+      end
+
+      x_config = app.config.x.try(:recording_studio_orderable)
+      RecordingStudioOrderable.configuration.merge!(x_config.to_h) if x_config.respond_to?(:to_h)
+    rescue StandardError
+      nil
+    end
+
+    initializer "recording_studio_orderable.integrate_recording_studio" do
+      config.to_prepare do
+        next unless defined?(RecordingStudio)
+
+        RecordingStudio.register_recordable_type("RecordingStudio::RecordingOrder")
+
+        next unless defined?(RecordingStudio::Recording)
+
+        unless RecordingStudio::Recording.included_modules.include?(RecordingStudioOrderable::RecordingExtensions)
+          RecordingStudio::Recording.include(RecordingStudioOrderable::RecordingExtensions)
+        end
+      end
+    end
+  end
+end