+++ test/dummy/app/models/folder.rb
@@ -1,0 +1,5 @@
+class Folder < ApplicationRecord
+  include RecordingStudioOrderable::OrderableRecordable
+
+  recording_studio_order_group :pages, allows: [ "Page" ]
+end