+++ test/dummy/db/migrate/20260217072941_create_folders.rb
@@ -1,0 +1,9 @@
+class CreateFolders < ActiveRecord::Migration[8.1]
+  def change
+    create_table :folders, id: :uuid do |t|
+      t.string :name, null: false
+
+      t.timestamps
+    end
+  end
+end