+++ test/dummy/db/migrate/20260217072942_create_pages.rb
@@ -1,0 +1,9 @@
+class CreatePages < ActiveRecord::Migration[8.1]
+  def change
+    create_table :pages, id: :uuid do |t|
+      t.string :title, null: false
+
+      t.timestamps
+    end
+  end
+end